start using useShell for announcements
