+++ Composer/packages/client/src/pages/language-generation/table-view.tsx
@@ -17,12 +17,12 @@ import formatMessage from 'format-message';
 import { NeutralColors, FontSizes } from '@uifabric/fluent-theme';
 import { RouteComponentProps } from '@reach/router';
 import { LgTemplate } from '@bfc/shared';
-import { Async } from 'office-ui-fabric-react/lib/Utilities';
 
 import { StoreContext } from '../../store';
 import { increaseNameUtilNotExist } from '../../utils/lgUtil';
 import { navigateTo } from '../../utils';
 import { actionButton, formCell } from '../language-understanding/styles';
+import { useShell } from '../../shell';
 
 interface TableViewProps extends RouteComponentProps<{}> {
   dialogId: string;
@@ -30,28 +30,20 @@ interface TableViewProps extends RouteComponentProps<{}> {
 
 const TableView: React.FC<TableViewProps> = props => {
   const { state, actions } = useContext(StoreContext);
+  const { api } = useShell();
+  const { announce } = api;
   const { dialogs, lgFiles, projectId, locale } = state;
   const { dialogId } = props;
   const file = lgFiles.find(({ id }) => id === `${dialogId}.${locale}`);
   const createLgTemplate = useRef(debounce(actions.createLgTemplate, 500)).current;
   const copyLgTemplate = useRef(debounce(actions.copyLgTemplate, 500)).current;
   const removeLgTemplate = useRef(debounce(actions.removeLgTemplate, 500)).current;
-  const setMessage = useRef(debounce(actions.setMessage, 500)).current;
   const [templates, setTemplates] = useState<LgTemplate[]>([]);
   const listRef = useRef(null);
 
   const activeDialog = dialogs.find(({ id }) => id === dialogId);
 
   const [focusedIndex, setFocusedIndex] = useState(0);
-
-  const _async = new Async();
-
-  const announce = (message: string) => {
-    setMessage(message);
-    _async.setTimeout(() => {
-      setMessage(undefined);
-    }, 2000);
-  };
 
   useEffect(() => {
     if (!file || isEmpty(file)) return;
