+++ Composer/packages/client/src/shell/useShell.ts
@@ -9,7 +9,7 @@ import get from 'lodash/get';
 import * as luUtil from '../utils/luUtil';
 import { updateRegExIntent } from '../utils/dialogUtil';
 import { StoreContext } from '../store';
-import { getDialogData, setDialogData, sanitizeDialogData } from '../utils';
+import { getDialogData, setDialogData, sanitizeDialogData, announce } from '../utils';
 import { OpenAlertModal, DialogStyle } from '../components/Modal';
 import { getFocusPath } from '../utils/navigation';
 import { isAbsHosted } from '../utils/envUtil';
@@ -174,6 +174,7 @@ export function useShell(source: EventSource): { api: ShellApi; data: ShellData 
     redo: actions.redo,
     addCoachMarkRef: actions.onboardingAddCoachMarkRef,
     updateUserSettings: actions.updateUserSettings,
+    announce,
   };
 
   const currentDialog = useMemo(() => dialogs.find(d => d.id === dialogId), [dialogs, dialogId]);
