+++ Composer/packages/client/src/utils/index.ts
@@ -4,3 +4,4 @@
 export * from './fileUtil';
 export * from './dialogUtil';
 export * from './navigation';
+export * from './a11y';
